example/jp/sandwich-shop/infrastructure/requirements.txt
aws-cdk-lib==2.72.0
constructs>=10.0.0
